+++ src/Akka.Analyzers/AK1000/MustNotUseNewKeywordOnActorsAnalyzer.cs
@@ -4,6 +4,7 @@
 // </copyright>
 // -----------------------------------------------------------------------
 
+using Akka.Analyzers.Context;
 using Akka.Analyzers.Context.Core;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
